# api-gateway-service/docs/swagger/swagger.yaml
definitions:
  entity.HTTPMessage:
    properties:
      body:
        type: string
      title:
        type: string
    type: object
  entity.HTTPResp:
    properties:
      data: {}
      message:
        $ref: '#/definitions/entity.HTTPMessage'
      metadata:
        $ref: '#/definitions/entity.Meta'
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
  entity.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entity.Meta:
    properties:
      error:
        $ref: '#/definitions/entity.MetaError'
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      status:
        type: string
      status_code:
        type: integer
      time_elapsed:
        type: string
      timestamp:
        type: string
    type: object
  entity.MetaError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  entity.Pagination:
    properties:
      current_elements:
        type: integer
      current_page:
        type: integer
      cursor_end:
        type: string
      cursor_start:
        type: string
      sort_by:
        items:
          type: string
        type: array
      total_elements:
        type: integer
      total_pages:
        type: integer
    type: object
  entity.Ping:
    properties:
      message:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
  entity.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  entity.TriggerSchedulerParams:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/admin/scheduler/trigger:
    post:
      description: Trigger Scheduler
      parameters:
      - description: Parameter for triggering scheduler
        in: body
        name: trigger_input
        required: true
        schema:
          $ref: '#/definitions/entity.TriggerSchedulerParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.HTTPResp'
      security:
      - BearerAuth: []
      - XDateTimes: []
      summary: Trigger Scheduler
      tags:
      - Scheduler
  /auth/v1/login:
    post:
      description: This endpoint will sign in user with username and password
      parameters:
      - description: Input Username And Password
        in: body
        name: login_request
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.HTTPResp'
      summary: Sign In With Password
      tags:
      - Auth
  /auth/v1/register:
    post:
      description: This endpoint will register new user as member
      parameters:
      - description: Input Username And Password
        in: body
        name: register_request
        required: true
        schema:
          $ref: '#/definitions/entity.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.HTTPResp'
      summary: Register new user
      tags:
      - Auth
  /auth/v1/verify-email/{id}:
    post:
      description: This endpoint will mark user email as verified
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.HTTPResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.HTTPResp'
      summary: Verify user email
      tags:
      - Auth
  /ping:
    get:
      description: This endpoint will hit the server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Ping'
      summary: Health Check
      tags:
      - Server
swagger: "2.0"